Refactor common helper code to svg/helpers.js, add area nopes
