+++ modules/svg/index.js
@@ -9,13 +9,15 @@ export { svgLines } from './lines.js';
 export { svgMapillaryImages } from './mapillary_images.js';
 export { svgMapillarySigns } from './mapillary_signs.js';
 export { svgMidpoints } from './midpoints.js';
-export { svgOneWaySegments } from './one_way_segments.js';
+export { svgOneWaySegments } from './helpers.js';
 export { svgOpenstreetcamImages } from './openstreetcam_images.js';
 export { svgOsm } from './osm.js';
-export { svgPath } from './path.js';
-export { svgPointTransform } from './point_transform.js';
+export { svgPassiveVertex } from './helpers.js';
+export { svgPath } from './helpers.js';
+export { svgPointTransform } from './helpers.js';
 export { svgPoints } from './points.js';
-export { svgRelationMemberTags } from './relation_member_tags.js';
+export { svgRelationMemberTags } from './helpers.js';
+export { svgSegmentWay } from './helpers.js';
 export { svgTagClasses } from './tag_classes.js';
 export { svgTurns } from './turns.js';
 export { svgVertices } from './vertices.js';
