+++ modules/svg/lines.js
@@ -4,13 +4,13 @@ import _flatten from 'lodash-es/flatten';
 import _forOwn from 'lodash-es/forOwn';
 import _map from 'lodash-es/map';
 
-import { geoPath as d3_geoPath } from 'd3-geo';
 import { range as d3_range } from 'd3-array';
 
 import {
     svgOneWaySegments,
     svgPath,
     svgRelationMemberTags,
+    svgSegmentWay,
     svgTagClasses
 } from './index';
 
@@ -40,140 +40,20 @@ export function svgLines(projection, context) {
     function drawTargets(selection, graph, entities, filter) {
         var targetClass = context.getDebug('target') ? 'pink ' : 'nocolor ';
         var nopeClass = context.getDebug('target') ? 'red ' : 'nocolor ';
-        var getPath = svgPath(projection, graph).geojson;
-        // var getPath = d3_geoPath(projection);
-
+        var getPath = svgPath(projection).geojson;
         var activeID = context.activeID();
 
-        // Rather than drawing lines directly, we'll cut out pieces
-        // depending on which parts are active.
+        // The targets and nopes will be MultiLineString sub-segments of the ways
         var data = { targets: [], nopes: [] };
 
-        // Touch targets control which other vertices we can drag a vertex onto.
-        // - the activeID - nope
-        // - next to the activeID - yes (vertices will be merged)
-        // - 2 away from the activeID - nope (would create a self intersecting segment)
-        // - all others on a closed way - nope (would create a self intersecting polygon)
-        //
-        // 0 = active vertex - no touch/connect
-        // 1 = passive vertex - yes touch/connect
-        // 2 = adjacent vertex - special rules
-        function passive(d) {
-            if (!activeID) return 1;
-            if (activeID === d.id) return 0;
-
-            var parents = graph.parentWays(d);
-            var i, j;
-
-            for (i = 0; i < parents.length; i++) {
-                var nodes = parents[i].nodes;
-                var isClosed = parents[i].isClosed();
-                for (j = 0; j < nodes.length; j++) {   // find this vertex, look nearby
-                    if (nodes[j] === d.id) {
-                        var ix1 = j - 2;
-                        var ix2 = j - 1;
-                        var ix3 = j + 1;
-                        var ix4 = j + 2;
-
-                        if (isClosed) {  // wraparound if needed
-                            var max = nodes.length - 1;
-                            if (ix1 < 0)   ix1 = max + ix1;
-                            if (ix2 < 0)   ix2 = max + ix2;
-                            if (ix3 > max) ix3 = ix3 - max;
-                            if (ix4 > max) ix4 = ix4 - max;
-                        }
-
-                        if (nodes[ix1] === activeID) return 0;        // prevent self intersect
-                        else if (nodes[ix2] === activeID) return 2;   // adjacent - ok!
-                        else if (nodes[ix3] === activeID) return 2;   // adjacent - ok!
-                        else if (nodes[ix4] === activeID) return 0;   // prevent self intersect
-                        else if (isClosed && nodes.indexOf(activeID) !== -1) return 0;  // prevent self intersect
-                    }
-                }
-            }
-
-            return 1;
-        }
-
         entities.forEach(function(way) {
-            var coordGroups = { passive: [], active: [] };
-            var segment = [];
-            var startType = null;   // 0 = active, 1 = passive, 2 = adjacent
-            var currType = null;
-            var node;
-
-            for (var i = 0; i < way.nodes.length; i++) {
-
-                if (way.nodes[i] === activeID) {    // vertex is the activeID
-                    segment = [];                      // draw no segment here
-                    startType = null;
-                    continue;
-                }
-
-                node = graph.entity(way.nodes[i]);
-                currType = passive(node);
-
-                if (startType === null) {
-                    startType = currType;
-                }
-
-                if (currType !== startType) {    // line changes here - try to save a segment
-
-                    if (segment.length > 0) {         // finish previous segment
-                        segment.push(node.loc);
-
-                        if (startType === 2 || currType === 2) {            // one adjacent vertex
-                            coordGroups.active.push(segment);
-                        } else if (startType === 0 && currType === 0) {     // both active vertices
-                            coordGroups.active.push(segment);
-                        } else {
-                            coordGroups.passive.push(segment);
-                        }
-                    }
-
-                    segment = [];
-                    startType = currType;
-                }
-
-                segment.push(node.loc);
-            }
-
-            // complete whatever segment we ended on
-            if (segment.length > 1) {
-                if (startType === 2 || currType === 2) {            // one adjacent vertex
-                    coordGroups.active.push(segment);
-                } else if (startType === 0 && currType === 0) {     // both active vertices
-                    coordGroups.active.push(segment);
-                } else {
-                    coordGroups.passive.push(segment);
-                }
-            }
-
-            if (coordGroups.passive.length) {
-                data.targets.push({
-                    'type': 'Feature',
-                    'id': way.id,
-                    'geometry': {
-                        'type': 'MultiLineString',
-                        'coordinates': coordGroups.passive
-                    }
-                });
-            }
-
-            if (coordGroups.active.length) {
-                data.nopes.push({
-                    'type': 'Feature',
-                    'id': way.id + '-nope',   // break the ids on purpose
-                    'geometry': {
-                        'type': 'MultiLineString',
-                        'coordinates': coordGroups.active
-                    }
-                });
-            }
+            var features = svgSegmentWay(way, graph, activeID);
+            data.targets.push.apply(data.targets, features.passive);
+            data.nopes.push.apply(data.nopes, features.active);
         });
 
 
-        // Places to hover and connect
+        // Targets allow hover and vertex snapping
         var targets = selection.selectAll('.line.target-allowed')
             .filter(filter)
             .data(data.targets, function key(d) { return d.id; });
@@ -204,7 +84,6 @@ export function svgLines(projection, context) {
             .merge(nopes)
             .attr('d', getPath)
             .attr('class', function(d) { return 'way line target target-nope ' + nopeClass + d.id; });
-
     }
